Refactor message handler logging.
