+++ src/uk/ac/cam/ssjt2/dissertation/kdc/KDCMessageHandler.java
@@ -1,4 +1,4 @@
-package uk.ac.cam.ssjt2.dissertation.client;
+package uk.ac.cam.ssjt2.dissertation.kdc;
 
 import uk.ac.cam.ssjt2.dissertation.common.AuthenticationProtocol;
 import uk.ac.cam.ssjt2.dissertation.common.CipherTools;
@@ -6,7 +6,6 @@ import uk.ac.cam.ssjt2.dissertation.common.MessageHandlerBase;
 import uk.ac.cam.ssjt2.dissertation.common.messages.KDCRequestMessage;
 import uk.ac.cam.ssjt2.dissertation.common.messages.KDCResponseMessage;
 import uk.ac.cam.ssjt2.dissertation.common.messages.TestMessage;
-import uk.ac.cam.ssjt2.dissertation.kdc.AuthenticationKDC;
 
 import javax.crypto.SecretKey;
 import java.io.IOException;
@@ -28,17 +27,16 @@ public class KDCMessageHandler extends MessageHandlerBase {
     @Override
     public void handleMessage() throws IOException {
         byte header = (byte) m_InputStream.read();
-        System.out.println("[KDC] Received packet header: " + header);
+        log("Received packet header: " + header);
 
         switch(header) {
             case AuthenticationProtocol.HEADER_TEST:
-                System.out.println("[KDC] Received test message.");
-                m_KDC.UnhandledMessages.add(new TestMessage().readFromStream(m_InputStream));
+                log("Received test message.");
                 break;
             case AuthenticationProtocol.HEADER_KDC_REQUEST:
-                System.out.println("[KDC] Received KDC request message.");
+                log("Received KDC request message.");
                 KDCRequestMessage message = KDCRequestMessage.readFromStream(m_InputStream);
-                System.out.println("[KDC] KDC request from client " + message.getClientId() + " for target " + message.getTargetId() + " with nonce " + message.getClientNonce());
+                log("KDC request from client " + message.getClientId() + " for target " + message.getTargetId() + " with nonce " + message.getClientNonce());
 
                 SecretKey clientKey = m_KDC.getKey(message.getClientId());
                 SecretKey targetKey = m_KDC.getKey(message.getTargetId());
@@ -47,13 +45,21 @@ public class KDCMessageHandler extends MessageHandlerBase {
                     KDCResponseMessage response = new KDCResponseMessage(message.getClientId(), clientKey, message.getTargetId(), targetKey, message.getClientNonce(), sessionKey);
                     m_OutputStream.write(response.getBytes());
                 } catch (Exception e) {
-                    System.err.println("[KDC] Error occurred while forming KDC response to " + message.getClientId());
+                    logError("Error occurred while forming KDC response to " + message.getClientId());
                     e.printStackTrace();
                 }
                 break;
             default:
-                //throw new IllegalArgumentException("Unrecognized message header: " + header);
+                throw new IllegalArgumentException("Unrecognized message header: " + header);
         }
+    }
+
+    public void log(String message) {
+        System.out.println("[KDC] " + message);
+    }
+
+    public void logError(String message) {
+        System.err.println("[KDC] " + message);
     }
 
 }
